🔥(layouts) remove PostShare component from post footer
The PostShare component included facilities to share stuff on
social networks (facebook, google+, twitter). In our opinion such
links and logos are cheap cruft and not the kind we want to be
displaying here. Drop the component and remove a CSS adjustment that
was only needed to make space for it.